backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "An interactive English learning app that uses a roulette game mechanic to help users practice vocabulary and grammar."
authors = [
    {name = "SmokingA1",email = "denisshokolad1221@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
